backend/requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
dj-rest-auth==7.0.1
Django==5.2.9
django-allauth==65.13.1
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==21.2.0
idna==3.11
outscraper==6.0.2
Pillow==11.1.0
PyJWT==2.10.1
python-dotenv==1.2.1
qrcode==8.1.0
requests==2.32.5
sqlparse==0.5.4
urllib3==2.5.0
whitenoise==6.11.0